fix: --device_ratios rejected fractional values
Parsing the ratios with int() dated back to before AutoSplit, which
feeds learned float ratios through the very same consumers
(_tile_split_devices, quantize_linears_parallel, calibration_row_shards
-- the last one already documents that ratios may be floats). So the
only thing standing between a user and "-dr 1.5,20" was the CLI parse,
which died with "invalid literal for int() with base 10: '1.5'".

+++ exllamav3/conversion/convert_model.py
@@ -1019,7 +1019,7 @@ def main(args, job_state):
     devices = [int(d) for d in args["devices"].split(",")]
     device = torch.device(devices[0])
     if args.get("device_ratios"):
-        device_ratios = [int(d) for d in args["device_ratios"].split(",")]
+        device_ratios = [float(d) for d in args["device_ratios"].split(",")]
         assert len(devices) == len(device_ratios), "--devices and --device_ratios must be same length"
     else:
         device_ratios = None
